5.3

+++ src/Illuminate/Queue/DatabaseQueue.php
@@ -6,7 +6,6 @@ use DateTime;
 use Carbon\Carbon;
 use Illuminate\Database\Connection;
 use Illuminate\Queue\Jobs\DatabaseJob;
-use Illuminate\Database\Query\Expression;
 use Illuminate\Contracts\Queue\Queue as QueueContract;
 
 class DatabaseQueue extends Queue implements QueueContract
@@ -54,6 +53,19 @@ class DatabaseQueue extends Queue implements QueueContract
         $this->expire = $expire;
         $this->default = $default;
         $this->database = $database;
+    }
+
+    /**
+     * Get the size of the queue.
+     *
+     * @param  string  $queue
+     * @return int
+     */
+    public function size($queue = null)
+    {
+        return $this->database->table($this->table)
+                    ->where('queue', $this->getQueue($queue))
+                    ->count();
     }
 
     /**
@@ -160,12 +172,10 @@ class DatabaseQueue extends Queue implements QueueContract
     {
         $queue = $this->getQueue($queue);
 
-        if (! is_null($this->expire)) {
-            $this->releaseJobsThatHaveBeenReservedTooLong($queue);
-        }
+        $this->database->beginTransaction();
 
         if ($job = $this->getNextAvailableJob($queue)) {
-            $this->markJobAsReserved($job->id);
+            $job = $this->markJobAsReserved($job);
 
             $this->database->commit();
 
@@ -178,27 +188,6 @@ class DatabaseQueue extends Queue implements QueueContract
     }
 
     /**
-     * Release the jobs that have been reserved for too long.
-     *
-     * @param  string  $queue
-     * @return void
-     */
-    protected function releaseJobsThatHaveBeenReservedTooLong($queue)
-    {
-        $expired = Carbon::now()->subSeconds($this->expire)->getTimestamp();
-
-        $this->database->table($this->table)
-                    ->where('queue', $this->getQueue($queue))
-                    ->where('reserved', 1)
-                    ->where('reserved_at', '<=', $expired)
-                    ->update([
-                        'reserved' => 0,
-                        'reserved_at' => null,
-                        'attempts' => new Expression('attempts + 1'),
-                    ]);
-    }
-
-    /**
      * Get the next available job for the queue.
      *
      * @param  string|null  $queue
@@ -206,13 +195,13 @@ class DatabaseQueue extends Queue implements QueueContract
      */
     protected function getNextAvailableJob($queue)
     {
-        $this->database->beginTransaction();
-
         $job = $this->database->table($this->table)
                     ->lockForUpdate()
                     ->where('queue', $this->getQueue($queue))
-                    ->where('reserved', 0)
-                    ->where('available_at', '<=', $this->getTime())
+                    ->where(function ($query) {
+                        $this->isAvailable($query);
+                        $this->isReservedButExpired($query);
+                    })
                     ->orderBy('id', 'asc')
                     ->first();
 
@@ -220,16 +209,51 @@ class DatabaseQueue extends Queue implements QueueContract
     }
 
     /**
-     * Mark the given job ID as reserved.
+     * Modify the query to check for available jobs.
      *
-     * @param  string  $id
+     * @param  \Illuminate\Database\Query\Builder  $query
      * @return void
      */
-    protected function markJobAsReserved($id)
+    protected function isAvailable($query)
     {
-        $this->database->table($this->table)->where('id', $id)->update([
-            'reserved' => 1, 'reserved_at' => $this->getTime(),
+        $query->where(function ($query) {
+            $query->whereNull('reserved_at');
+            $query->where('available_at', '<=', $this->getTime());
+        });
+    }
+
+    /**
+     * Modify the query to check for jobs that are reserved but have expired.
+     *
+     * @param  \Illuminate\Database\Query\Builder  $query
+     * @return void
+     */
+    protected function isReservedButExpired($query)
+    {
+        $expiration = Carbon::now()->subSeconds($this->expire)->getTimestamp();
+
+        $query->orWhere(function ($query) use ($expiration) {
+            $query->where('reserved_at', '<=', $expiration);
+        });
+    }
+
+    /**
+     * Mark the given job ID as reserved.
+     *
+     * @param \stdClass $job
+     * @return \stdClass
+     */
+    protected function markJobAsReserved($job)
+    {
+        $job->attempts = $job->attempts + 1;
+        $job->reserved_at = $this->getTime();
+
+        $this->database->table($this->table)->where('id', $job->id)->update([
+            'reserved_at' => $job->reserved_at,
+            'attempts' => $job->attempts,
         ]);
+
+        return $job;
     }
 
     /**
@@ -241,7 +265,13 @@ class DatabaseQueue extends Queue implements QueueContract
      */
     public function deleteReserved($queue, $id)
     {
-        $this->database->table($this->table)->where('id', $id)->delete();
+        $this->database->beginTransaction();
+
+        if ($this->database->table($this->table)->lockForUpdate()->find($id)) {
+            $this->database->table($this->table)->where('id', $id)->delete();
+        }
+
+        $this->database->commit();
     }
 
     /**
@@ -270,12 +300,11 @@ class DatabaseQueue extends Queue implements QueueContract
     {
         return [
             'queue' => $queue,
-            'payload' => $payload,
             'attempts' => $attempts,
-            'reserved' => 0,
             'reserved_at' => null,
             'available_at' => $availableAt,
             'created_at' => $this->getTime(),
+            'payload' => $payload,
         ];
     }
 
@@ -298,26 +327,5 @@ class DatabaseQueue extends Queue implements QueueContract
     public function getDatabase()
     {
         return $this->database;
-    }
-
-    /**
-     * Get the expiration time in seconds.
-     *
-     * @return int|null
-     */
-    public function getExpire()
-    {
-        return $this->expire;
-    }
-
-    /**
-     * Set the expiration time in seconds.
-     *
-     * @param  int|null  $seconds
-     * @return void
-     */
-    public function setExpire($seconds)
-    {
-        $this->expire = $seconds;
     }
 }
